break main RichSuggestBox.cs into smaller files
